moving test utilities to commons
