+++ feature-dogs/src/test/java/feature/dogs/data/repository/DogsApiRepositoryImplTest.kt
@@ -1,5 +1,6 @@
 package feature.dogs.data.repository
 
+import feature.commons.test_utils.CoroutineTestRule
 import feature.dogs.data.remoteDataSource.RemoteDataSource
 import feature.dogs.domain.model.BreedDomain
 import feature.dogs.domain.model.HeightDomain
@@ -10,7 +11,6 @@ import feature.dogs.presentation.model.BreedPresentation
 import feature.dogs.presentation.model.HeightPresentation
 import feature.dogs.presentation.model.ImagePresentation
 import feature.dogs.presentation.model.WeightPresentation
-import feature.dogs.test_utils.CoroutineTestRule
 import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.mockk
